pyproject.toml
[project]
# https://packaging.python.org/en/latest/specifications/declaring-project-metadata/
name = "datalinker"
description = "Service and data discovery for Rubin Science Platform"
license = "MIT"
license-files = ["LICENSE"]
readme = "README.md"
keywords = ["rubin", "lsst"]
# https://pypi.org/classifiers/
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Natural Language :: English",
    "Operating System :: POSIX",
    "Private :: Do Not Upload",
    "Typing :: Typed",
]
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.100",
    "httpx",
    "jinja2",
    "lsst-daf-butler[remote]",
    "pydantic>=2.11",
    "pydantic-settings",
    "pyyaml",
    "safir>=6.2.0",
    "structlog",
    "uvicorn[standard]",
]
dynamic = ["version"]

[[project.authors]]
name = "Association of Universities for Research in Astronomy, Inc. (AURA)"
email = "sqre-admin@lists.lsst.org"

[project.urls]
Homepage = "https://github.com/lsst-sqre/datalinker"
Source = "https://github.com/lsst-sqre/datalinker"
"Change log" = "https://github.com/lsst-sqre/datalinker/blob/main/CHANGELOG.md"
"Issue tracker" = "https://github.com/lsst-sqre/datalinker/issues"

[build-system]
requires = [
    "setuptools>=61",
    "wheel",
    "setuptools_scm[toml]>=6.2"
]
build-backend = 'setuptools.build_meta'

[dependency-groups]
dev = [
    "asgi-lifespan",
    "coverage[toml]",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-sugar",
    "respx",
    "scriv[toml]",
]
lint = [
    "pre-commit",
    "pre-commit-uv",
    "ruff>=0.9",
]
tox = [
    "tox>=4.24",
    "tox-uv>=1.25",
]
typing = [
    "mypy>=1.15",
    "types-PyYAML",
]

[tool.coverage.run]
parallel = true
branch = true
source = ["datalinker"]

[tool.coverage.paths]
source = ["src", ".tox/*/site-packages"]

[tool.coverage.report]
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:"
]

[tool.mypy]
disallow_untyped_defs = true
disallow_incomplete_defs = true
ignore_missing_imports = true
local_partial_types = true
plugins = ["pydantic.mypy"]
no_implicit_reexport = true
show_error_codes = true
strict_equality = true
warn_redundant_casts = true
warn_unreachable = true
warn_unused_ignores = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true
warn_required_dynamic_aliases = true
warn_untyped_fields = true

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
asyncio_mode = "strict"
# The python_files setting is not for test detection (pytest will pick up any
# test files named *_test.py without this setting) but to enable special
# assert processing in any non-test supporting files under tests.  We
# conventionally put test support functions under tests.support and may
# sometimes use assert in test fixtures in conftest.py, and pytest only
# enables magical assert processing (showing a full diff on assert failures
# with complex data structures rather than only the assert message) in files
# listed in python_files.
python_files = ["tests/*.py", "tests/*/*.py"]

# Use the generic Ruff configuration in ruff.toml and extend it with only
# project-specific settings. Add a [tool.ruff.lint.extend-per-file-ignores]
# section for project-specific ignore rules.
[tool.ruff]
extend = "ruff-shared.toml"

[tool.ruff.lint.extend-per-file-ignores]
"src/datalinker/handlers/**" = [
    "S608",    # this is ADQL, and the parameters are validated
]
"tests/support/butler.py" = [
    "N802",    # mocking a method that doesn't use snake case
]

[tool.ruff.lint.isort]
known-first-party = ["datalinker", "tests"]
split-on-trailing-comma = false

[tool.scriv]
categories = [
    "Backwards-incompatible changes",
    "New features",
    "Bug fixes",
    "Other changes",
]
entry_title_template = "{{ version }} ({{ date.strftime('%Y-%m-%d') }})"
format = "md"
md_header_level = "2"
new_fragment_template = "file:changelog.d/_template.md.jinja"
skip_fragments = "_template.md.jinja"

[tool.setuptools_scm]
